Changed shared memory allocation to create the file using mkstemp rather than a filename based the USER environement var

+++ src/mpid/ch3/channels/nemesis/nemesis/src/mpid_nem_alloc.c
@@ -5,34 +5,86 @@
 
 void MPID_nem_seg_create(MPID_nem_seg_ptr_t memory, int size, int num_local, int local_rank)
 {
-    char name[MPID_NEM_MAX_FNAME_LEN]  = "/tmp/shmem.map_";
-    char *user           = getenv("USER");
     int ret;
-    int errno;
+
+    memory->max_size = size;
     
-    strcat (name, user);
-    strncpy (memory->file_name, name, MPID_NEM_MAX_FNAME_LEN);
-    memory->max_size     = size ;
-    memory->base_descs   = open (memory->file_name, O_RDWR | O_CREAT, S_IRWXU);
-    if (memory->base_descs == -1)
+    if (local_rank == 0)
     {
-	perror ("Error opening shared file.");
-	exit (-1);
+	char filename[MPID_NEM_MAX_FNAME_LEN] = "/dev/shm/nemesis_shar_tmpXXXXXX";
+	memory->base_descs = mkstemp (filename);
+	if (memory->base_descs == -1)
+	{
+	    fprintf (stderr, "%d: Error opening shared file \"%s\": %s \n", local_rank, memory->file_name, strerror(errno));
+	    exit (-1);
+	}
+	MPIU_Strncpy (memory->file_name, filename, MPID_NEM_MAX_FNAME_LEN);
+
+	ret = ftruncate(memory->base_descs, memory->max_size);
+	if (ret == -1)
+	{
+	    fprintf (stderr, "%d: Error resizing shared file to %d: %s \n", local_rank, memory->max_size, strerror(errno));
+	    unlink (memory->file_name);
+	    exit (-1);
+	}
+	
+	/* post name of shared file */
+	memset (pmi_val, 0, pmi_val_max_sz);
+	snprintf (pmi_val, pmi_val_max_sz, "%s", memory->file_name);
+	memset (pmi_key, 0, pmi_key_max_sz);
+	snprintf (pmi_key, pmi_key_max_sz, "sharedFilename");
+	ret = PMI_KVS_Put (pmi_kvs_name, pmi_key, pmi_val);
+	if (ret != 0)
+	{
+	    unlink (memory->file_name);
+	    FATAL_ERROR ("PMI_KVS_Put failed %d", ret);
+	}
+	ret = PMI_KVS_Commit (pmi_kvs_name);
+	if (ret != 0)
+	{
+	    unlink (memory->file_name);
+	    FATAL_ERROR ("PMI_KVS_Commit failed %d", ret);
+	}
+
+	ret = PMI_Barrier();
+	if (ret != 0)
+	{
+	    unlink (memory->file_name);
+	    FATAL_ERROR ("PMI_Barrier failed %d", ret);
+	}
     }
-    
-    ret = ftruncate(memory->base_descs, memory->max_size);
-    if (ret == -1)
+    else
     {
-	perror ("Resizing shared file failed.");
-	printf ("max size = %d\n", memory->max_size);
-	unlink (memory->file_name);
-	exit (-1);
+	ret = PMI_Barrier();
+	if (ret != 0)
+	{
+	    unlink (memory->file_name);
+	    FATAL_ERROR ("PMI_Barrier failed %d", ret);
+	}
+
+	/* get name of shared file */
+	snprintf (pmi_key, pmi_key_max_sz, "sharedFilename");
+	memset (pmi_val, 0, pmi_val_max_sz);
+	ret = PMI_KVS_Get (pmi_kvs_name, pmi_key, pmi_val, pmi_val_max_sz);
+	if (ret != 0)
+	{
+	    FATAL_ERROR ("PMI_KVS_Get failed %d", ret);
+	}
+
+	MPIU_Strncpy (memory->file_name, pmi_val, MPID_NEM_MAX_FNAME_LEN);
+
+	memory->base_descs = open (memory->file_name, O_RDWR);
+	if (memory->base_descs == -1)
+	{
+	    fprintf (stderr, "%d: Error opening shared file \"%s\": %s \n", local_rank, memory->file_name, strerror(ret));
+	    exit (-1);
+	}
     }
     
     memory->base_addr = mmap (NULL, memory->max_size, PROT_READ | PROT_WRITE, MAP_SHARED, memory->base_descs, 0);
     if (memory->base_addr == MAP_FAILED)
     {
-	perror ("Error mmap()ing shared memory region.");
+	fprintf (stderr, "%d: Error mmap()ing shared memory region: %s \n", local_rank, strerror(errno));
 	unlink (memory->file_name);
 	exit (-1);
     }
@@ -43,15 +95,15 @@ void MPID_nem_seg_create(MPID_nem_seg_ptr_t memory, int size, int num_local, int
     }
     while (ret == EINTR);
     if (ret)
-	perror ("Error closing shared memory file.");
+	fprintf (stderr, "%d: Error closing shared memory file: %s \n", local_rank, strerror(errno));
     
     memset (memory->base_addr, 0, memory->max_size);
     
-    errno = PMI_Barrier();
-    if (errno != 0)
+    ret = PMI_Barrier();
+    if (ret != 0)
     {
 	unlink (memory->file_name);
-	FATAL_ERROR ("PMI_Barrier failed %d", errno);
+	FATAL_ERROR ("PMI_Barrier failed %d", ret);
     }
     
     unlink (memory->file_name);
@@ -83,20 +135,20 @@ void MPID_nem_check_alloc (int num_processes)
     int rank    = MPID_nem_mem_region.local_rank;
     int address = 0;
     int base, found, index;
-    int errno;
+    int ret;
     
-    errno = PMI_Barrier();
-    if (errno != 0)
-	FATAL_ERROR ("PMI_Barrier failed %d", errno);
+    ret = PMI_Barrier();
+    if (ret != 0)
+	FATAL_ERROR ("PMI_Barrier failed %d", ret);
 
     address =  ((MPID_nem_addr_t)(MPID_nem_mem_region.memory.current_addr));
     MPID_NEM_MEMCPY(&(((int*)(MPID_nem_mem_region.memory.current_addr))[rank]),
 	   &address,
 	   sizeof(MPID_nem_addr_t));
 
-    errno = PMI_Barrier();
-    if (errno != 0)
-	FATAL_ERROR ("PMI_Barrier failed %d", errno);
+    ret = PMI_Barrier();
+    if (ret != 0)
+	FATAL_ERROR ("PMI_Barrier failed %d", ret);
 
     base  = ((int *)MPID_nem_mem_region.memory.current_addr)[rank];
     found = 1 ;
@@ -110,9 +162,9 @@ void MPID_nem_check_alloc (int num_processes)
 	      }		  
 	  }	     
       }
-    errno = PMI_Barrier();
-    if (errno != 0)
-	FATAL_ERROR ("PMI_Barrier failed %d", errno);
+    ret = PMI_Barrier();
+    if (ret != 0)
+	FATAL_ERROR ("PMI_Barrier failed %d", ret);
 
     if (found == num_processes)
       {
